--- app/src/main/java/com/example/ppt_kotders/SavedPreferencesUser.kt
@@ -1,44 +1,0 @@
-package com.example.ppt_kotders
-
-import android.content.Context
-import android.content.SharedPreferences
-import android.preference.PreferenceManager
-
-object SavedPreferencesUser {
-    const val EMAIL= "email"
-    const val USERNAME="username"
-
-    private  fun getSharedPreference(ctx: Context?): SharedPreferences? {
-        return PreferenceManager.getDefaultSharedPreferences(ctx)
-    }
-
-    private fun  editor(context: Context, const:String, string: String){
-        getSharedPreference(
-            context
-        )?.edit()?.putString(const,string)?.apply()
-    }
-
-    fun getEmail(context: Context)= getSharedPreference(
-        context
-    )?.getString(EMAIL,"")
-
-    fun setEmail(context: Context, email: String){
-        editor(
-            context,
-            EMAIL,
-            email
-        )
-    }
-
-    fun setUsername(context: Context, username:String){
-        editor(
-            context,
-            USERNAME,
-            username
-        )
-    }
-
-    fun getUsername(context: Context) = getSharedPreference(
-        context
-    )?.getString(USERNAME,"")
-}
